third_party/WebKit/LayoutTests/inspector/timeline/timeline-auto-record.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script>

function test()
{
    var panel = WebInspector.panel("timeline");
    var model = panel._model;
    InspectorTest.assertTrue(!model._clientInitiatedRecording, "timeline is being recorded");
    InspectorTest.assertTrue(!model.records().length, "timeline has already been recorded");

    WebInspector.showPanel("console");
    InspectorTest.runWhenPageLoads(step1);
    WebInspector.resourceTreeModel.reloadPage();

    function step1()
    {
        InspectorTest.assertTrue(!model._clientInitiatedRecording, "timeline started recording");
        InspectorTest.assertTrue(!model.records().length, "timeline has been recorded");

        WebInspector.showPanel("timeline");
        InspectorTest.runWhenPageLoads(step2);
        WebInspector.resourceTreeModel.reloadPage();
    }

    function step2()
    {
        InspectorTest.assertTrue(model.records().length, "timeline should have been recorded");
        InspectorTest.assertTrue(!model._clientInitiatedRecording, "timeline recording should have finished");
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>Test auto-recording of Timeline.</p>

</body>
</html>
