third_party/WebKit/LayoutTests/inspector/timeline/timeline-auto-record-expected.txt

Test auto-recording of Timeline.

Reloading page on console panel
Page reloaded.
Reloading page on timeline panel
Recording started
Initiated by user: true
Page reloaded.
Recording stopped
